requirements.txt
easyocr
pillow
textblob
torch
torchvision
git+https://github.com/openai/CLIP.git 
diffusers
transformers